user separation to a domain layer
