+++ Domain/Class1.cs
@@ -1,0 +1,14 @@
+﻿using System;
+
+namespace Domain
+{
+    public class User
+    {
+        public string Name { get; set; }
+        public string Email { get; set; }
+        public string Address { get; set; }
+        public string Phone { get; set; }
+        public string UserType { get; set; }
+        public decimal Money { get; set; }
+    }
+}
