+++ Sat.Recruitment.Api/Controllers/Result.cs
@@ -1,0 +1,8 @@
+﻿namespace Sat.Recruitment.Api.Controllers
+{
+    public class Result
+    {
+        public bool IsSuccess { get; set; }
+        public string Errors { get; set; }
+    }
+}
